cleanup: remove settings v1
