--- ui/desktop/src/components/settings/api_keys/types.ts
@@ -1,13 +1,0 @@
-export interface ProviderResponse {
-  supported: boolean;
-  name?: string;
-  description?: string;
-  models?: string[];
-  config_status: Record<string, ConfigDetails>;
-}
-
-export interface ConfigDetails {
-  key: string;
-  is_set: boolean;
-  location?: string;
-}
