--- ui/desktop/src/components/settings/types.ts
@@ -1,13 +1,0 @@
-import { FullExtensionConfig } from '../../extensions';
-
-export interface Model {
-  id: string;
-  name: string;
-  description: string;
-  enabled: boolean;
-}
-
-export interface Settings {
-  models: Model[];
-  extensions: FullExtensionConfig[];
-}
