+++ ui/desktop/src/renderer.tsx
@@ -1,9 +1,7 @@
 import React, { Suspense, lazy } from 'react';
 import ReactDOM from 'react-dom/client';
-import { ModelProvider } from './components/settings/models/ModelContext';
 import { ConfigProvider } from './components/ConfigContext';
 import { ErrorBoundary } from './components/ErrorBoundary';
-import { ActiveKeysProvider } from './components/settings/api_keys/ActiveKeysContext';
 import { patchConsoleLogging } from './utils';
 import SuspenseLoader from './suspense-loader';
 
@@ -15,13 +13,9 @@ ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <Suspense fallback={SuspenseLoader()}>
       <ConfigProvider>
-        <ModelProvider>
-          <ActiveKeysProvider>
-            <ErrorBoundary>
-              <App />
-            </ErrorBoundary>
-          </ActiveKeysProvider>
-        </ModelProvider>
+        <ErrorBoundary>
+          <App />
+        </ErrorBoundary>
       </ConfigProvider>
     </Suspense>
   </React.StrictMode>
